refactor: extract PDF note matching
